Projects/T30_Filters_tool/requirements.txt
distlib==0.3.9
filelock==3.16.1
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
platformdirs==4.3.6
virtualenv==20.27.1
